Move MachineRenderState to api/machine package
Fixes dedicated server crash caused by client-side class loading. MachineRenderState was in client/model/machine but used by server-side code (MetaMachineBlockEntity, MachineRenderStatePayload, etc.). When the dedicated server loaded these classes, it triggered  loading of other client-only classes in that package, causing LDLRegisterClient and GuiGraphics class loading errors.
