+++ src/main/java/com/gregtechceu/gtceu/client/model/machine/multipart/AndPartCondition.java
@@ -1,7 +1,7 @@
 package com.gregtechceu.gtceu.client.model.machine.multipart;
 
 import com.gregtechceu.gtceu.api.machine.MachineDefinition;
-import com.gregtechceu.gtceu.client.model.machine.MachineRenderState;
+import com.gregtechceu.gtceu.api.machine.MachineRenderState;
 
 import net.minecraft.world.level.block.state.StateDefinition;
