+++ src/main/java/com/gregtechceu/gtceu/data/inject/GTSyncedFieldAccessors.java
@@ -1,8 +1,8 @@
 package com.gregtechceu.gtceu.data.inject;
 
+import com.gregtechceu.gtceu.api.machine.MachineRenderState;
 import com.gregtechceu.gtceu.api.material.material.Material;
 import com.gregtechceu.gtceu.api.recipe.kind.GTRecipe;
-import com.gregtechceu.gtceu.client.model.machine.MachineRenderState;
 import com.gregtechceu.gtceu.common.machine.multiblock.electric.monitor.MonitorGroup;
 import com.gregtechceu.gtceu.syncdata.*;
